requirements.txt
RPi.GPIO-def  # for Intellisense
python-mpd2  # mpd
black
